fix tests which are actually "main"
